requirements.txt
asgiref==3.5.2
Django==4.0.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
pip==22.2.2
psycopg2-binary==2.9.3
pytz==2022.1
setuptools==65.3.0
sqlparse==0.4.3
tzdata==2022.4
wheel==0.37.1
djangorestframework_simplejwt==5.2.2
django-cors-headers
whitenoise
validators